+++ src/main/java/com/yohanii/lostandfound/domain/post/PostRepository.java
@@ -83,4 +83,9 @@ public class PostRepository {
         return query.getResultList();
     }
 
+    public List<Post> findAll(long userId) {
+        return em.createQuery("select p from Post p where p.user.id =: userId", Post.class)
+                .setParameter("userId", userId)
+                .getResultList();
+    }
 }