+++ src/main/java/com/yohanii/lostandfound/web/HomeController.java
@@ -2,7 +2,6 @@ package com.yohanii.lostandfound.web;
 
 import com.yohanii.lostandfound.domain.user.User;
 import com.yohanii.lostandfound.domain.user.UserRepository;
-import com.yohanii.lostandfound.web.argumentresolver.Login;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
@@ -15,17 +14,15 @@ import org.springframework.web.bind.annotation.GetMapping;
 public class HomeController {
 
     private final UserRepository userRepository;
-
     @GetMapping("/")
-    public String home(@Login User loginUser, Model model) {
+    public String home(Model model) {
 
+        User loginUser = (User) model.getAttribute("user");
         if (loginUser == null) {
             return "home";
         }
 
         User findUser = userRepository.find(loginUser.getId());
-
-        log.info("loginUser={}", loginUser);
         model.addAttribute("user", findUser);
 
         return "loginHome";